Refactor: rename cookieStore to syncCookieStore
