+++ packages/extension/src/view/stateProviders/syncCookieStore/index.tsx
